LightSchedule Refactor
Breaking apart monolithic file
